+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to ExcelExtract will be documented in this file.
 - Column names can contain tokens, generating dynamic columns.
 - Columns are implicit tokens, and can be used for calculations.
 - intrarow property is added to loops
+- basic table added
 
 ## [0.4.0] - 2025-05-05